Adds music preview.
Adds feature to preview musics using the systems players. Right now its hardcoded to use
amarok, althought work is under way to support various players.
The feature uses dbus to send commands to the player.

Everything seems to be working fine, but there's a problem with changing the icon when the
music is playing.
Another problem is to know when the music is over, maybe hook a callback to the player to
inform tagtool when the music changes, this way we can update our UI.